some code polishing done. testing
